+++ spec/features/togls_spec.rb
@@ -157,4 +157,23 @@ off - test3 - test3 readable description
 }).to_stdout
     end
   end
+
+  describe "defining feature toggles in additional registry" do
+    it "creates an isolated registred with a feature toggled off" do
+      Togls.features do
+        feature(:test, "some human readable description").on
+      end
+
+      klass = Class.new do
+        include Togls::ReleaseToggleRegistryManager
+      end
+
+      klass.features do
+        feature(:test, "some human readable description").off
+      end
+
+      expect(Togls.feature(:test).on?).to eq(true)
+      expect(klass.feature(:test).on?).to eq(false)
+    end
+  end
 end